requirements.txt
# Core
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.3

# Database
psycopg2-binary==2.9.9
supabase==1.1.1

# AI/LLM
langchain-together==0.1.4
langchain-core==0.1.24
langgraph==0.0.33
langchain-community==0.0.20

# Utilities
python-dotenv==1.0.0
typing-extensions==4.9.0
requests==2.31.0
httpx==0.27.2
# Development
pytest==7.4.3
black==23.12.1
mypy==1.8.0